Move tests to subfolders. Ref #10
